navarra/emergency-ai-agent/requirements.txt
agent
google-adk
pyttsx3
google-agents
# google-generativeai
SpeechRecognition
pyaudio
pipwin
google-generativeai[adk]